+++ test/int/sql/test_alter_index.py
@@ -1,0 +1,9 @@
+from framework.sqltester import (
+    ClusterTwoInstances,
+    sql_test_file,
+)
+
+
+@sql_test_file("alter_index_rename.sql")
+class TestRenameIndex(ClusterTwoInstances):
+    pass